Update and rename slps3.js to s1.js

+++ s1.js
@@ -141,7 +141,7 @@ async function getGnana(){
 
   bananaCv = parseInt(poolSplit,0);
   gnanaCv = parseInt(poolGnana,0);
-  bananaMath = (bananaCv*bananaUsdPrice).toString();
+  bananaMath = (BigInt(bananaCv*bananaUsdPrice)).toString();
   gnanaMath = gnanaCv*gnanaPrice;
 
   console.log(bananaMath.slice(0,-18));
